Rename component file name to match guide
- Renam onboarding component to `on-boarding` to match the style-guide.
