+++ client/src/components/on-boarding.tsx
